Refactor into two apps: api and lti
