+++ api/tests/test_utils.py
@@ -1,354 +1,19 @@
 import os
-import time
 from datetime import date, datetime, timedelta
 
 import oauth2
-import oauthlib.oauth1.rfc5849.signature as oauth_signature
 from django.conf import settings
 from django.core.exceptions import PermissionDenied
-from django.shortcuts import reverse
-from django.test import RequestFactory, TestCase, override_settings
+from django.test import TestCase
 from wimsapi import Class, User, WimsAPI
 
-from wims import utils
-from wims.exceptions import BadRequestException
-from wims.models import LMS, WIMS, WimsClass, WimsUser
+from api import utils
+from api.models import LMS, WIMS, WimsClass, WimsUser
 
-
-FAKE_CREDENTIALS = {
-    'provider1': 'secret1',
-}
+from lti_app.utils import parse_parameters
 
 # URL to the WIMS server used for tests, the server must recogned ident 'myself' and passwd 'toto'
 WIMS_URL = os.getenv("WIMS_URL") or "http://localhost:7777/wims/wims.cgi"
-
-
-
-@override_settings(LTI_OAUTH_CREDENTIALS=FAKE_CREDENTIALS)
-class IsValidRequestTestCase(TestCase):
-    
-    def test_is_valid_request_ok(self):
-        params = {
-            'lti_message_type':                   'basic-lti-launch-request',
-            'lti_version':                        'LTI-1p0',
-            'launch_presentation_locale':         'fr-FR',
-            'resource_link_id':                   'X',
-            'context_id':                         'X',
-            'context_title':                      "A title",
-            'user_id':                            'X',
-            'lis_person_contact_email_primary':   'X',
-            'lis_person_name_family':             'X',
-            'lis_person_name_given':              'X',
-            'tool_consumer_instance_description': 'X',
-            'tool_consumer_instance_guid':        'elearning.u-pem.fr',
-            'oauth_consumer_key':                 'provider1',
-            'oauth_signature_method':             'HMAC-SHA1',
-            'oauth_timestamp':                    str(oauth2.generate_timestamp()),
-            'oauth_nonce':                        oauth2.generate_nonce(),
-            'roles':                              "Learner"
-        }
-        
-        norm_params = oauth_signature.normalize_parameters([(k, v) for k, v in params.items()])
-        # Last 'dns' is the arg use in request factory
-        uri = oauth_signature.normalize_base_string_uri("https://testserver/wims/1/")
-        base_string = oauth_signature.construct_base_string("POST", uri, norm_params)
-        
-        params['oauth_signature'] = oauth_signature.sign_hmac_sha1(base_string, "secret1", None)
-        request = RequestFactory().post(reverse("wims:wims_class", args=[1]), secure=True)
-        request.POST = params
-        
-        utils.is_valid_request(request)
-    
-    
-    def test_is_valid_request_wrong_lti_message_type(self):
-        params = {
-            'lti_message_type':                   'wrong',
-            'lti_version':                        'LTI-1p0',
-            'launch_presentation_locale':         'fr-FR',
-            'resource_link_id':                   'X',
-            'context_id':                         'X',
-            'context_title':                      "A title",
-            'user_id':                            'X',
-            'lis_person_contact_email_primary':   'X',
-            'lis_person_name_family':             'X',
-            'lis_person_name_given':              'X',
-            'tool_consumer_instance_description': 'X',
-            'tool_consumer_instance_guid':        'elearning.u-pem.fr',
-            'oauth_consumer_key':                 'provider1',
-            'oauth_signature_method':             'HMAC-SHA1',
-            'oauth_timestamp':                    str(oauth2.generate_timestamp()),
-            'oauth_nonce':                        oauth2.generate_nonce(),
-            'roles':                              "Learner"
-        }
-        
-        norm_params = oauth_signature.normalize_parameters([(k, v) for k, v in params.items()])
-        # Last 'dns' is the arg use in request factory
-        uri = oauth_signature.normalize_base_string_uri("https://testserver/wims/1/")
-        base_string = oauth_signature.construct_base_string("POST", uri, norm_params)
-        
-        params['oauth_signature'] = oauth_signature.sign_hmac_sha1(base_string, "secret1", None)
-        request = RequestFactory().post(reverse("wims:wims_class", args=[1]), secure=True)
-        request.POST = params
-        
-        with self.assertRaises(BadRequestException) as r:
-            utils.is_valid_request(request)
-        self.assertEqual("LTI request is invalid, parameter 'lti_message_type' must be equal to "
-                         "'basic-lti-launch-request'",
-                         str(r.exception))
-    
-    
-    @override_settings(LTI_OAUTH_CREDENTIALS=False)
-    def test_is_valid_request_no_lti_oauth_credentials(self):
-        params = {
-            'lti_message_type':                   'basic-lti-launch-request',
-            'lti_version':                        'LTI-1p0',
-            'launch_presentation_locale':         'fr-FR',
-            'resource_link_id':                   'X',
-            'context_id':                         'X',
-            'context_title':                      "A title",
-            'user_id':                            'X',
-            'lis_person_contact_email_primary':   'X',
-            'lis_person_name_family':             'X',
-            'lis_person_name_given':              'X',
-            'tool_consumer_instance_description': 'X',
-            'tool_consumer_instance_guid':        'elearning.u-pem.fr',
-            'oauth_consumer_key':                 'provider1',
-            'oauth_signature_method':             'HMAC-SHA1',
-            'oauth_timestamp':                    str(oauth2.generate_timestamp()),
-            'oauth_nonce':                        oauth2.generate_nonce(),
-            'roles':                              "Learner"
-        }
-        
-        norm_params = oauth_signature.normalize_parameters([(k, v) for k, v in params.items()])
-        # Last 'dns' is the arg use in request factory
-        uri = oauth_signature.normalize_base_string_uri("https://testserver/wims/1/")
-        base_string = oauth_signature.construct_base_string("POST", uri, norm_params)
-        
-        params['oauth_signature'] = oauth_signature.sign_hmac_sha1(base_string, "secret1", None)
-        request = RequestFactory().post(reverse("wims:wims_class", args=[1]), secure=True)
-        request.POST = params
-        
-        with self.assertRaises(BadRequestException) as r:
-            utils.is_valid_request(request)
-        self.assertEqual("Missing LTI_OAUTH_CREDENTIALS in settings.", str(r.exception))
-    
-    
-    def test_is_valid_request_invalid_key(self):
-        params = {
-            'lti_message_type':                   'basic-lti-launch-request',
-            'lti_version':                        'LTI-1p0',
-            'launch_presentation_locale':         'fr-FR',
-            'resource_link_id':                   'X',
-            'context_id':                         'X',
-            'context_title':                      "A title",
-            'user_id':                            'X',
-            'lis_person_contact_email_primary':   'X',
-            'lis_person_name_family':             'X',
-            'lis_person_name_given':              'X',
-            'tool_consumer_instance_description': 'X',
-            'tool_consumer_instance_guid':        'elearning.u-pem.fr',
-            'oauth_consumer_key':                 'provider1',
-            'oauth_signature_method':             'HMAC-SHA1',
-            'oauth_timestamp':                    str(oauth2.generate_timestamp()),
-            'oauth_nonce':                        oauth2.generate_nonce(),
-            'roles':                              "Learner"
-        }
-        
-        norm_params = oauth_signature.normalize_parameters([(k, v) for k, v in params.items()])
-        # Last 'dns' is the arg use in request factory
-        uri = oauth_signature.normalize_base_string_uri("https://testserver/wims/1/")
-        base_string = oauth_signature.construct_base_string("POST", uri, norm_params)
-        
-        params['oauth_signature'] = oauth_signature.sign_hmac_sha1(base_string, "secret2", None)
-        request = RequestFactory().post(reverse("wims:wims_class", args=[1]), secure=True)
-        request.POST = params
-        
-        with self.assertRaises(PermissionDenied):
-            utils.is_valid_request(request)
-    
-    
-    def test_is_valid_request_unknown_consumer(self):
-        params = {
-            'lti_message_type':                   'basic-lti-launch-request',
-            'lti_version':                        'LTI-1p0',
-            'launch_presentation_locale':         'fr-FR',
-            'resource_link_id':                   'X',
-            'context_id':                         'X',
-            'context_title':                      "A title",
-            'user_id':                            'X',
-            'lis_person_contact_email_primary':   'X',
-            'lis_person_name_family':             'X',
-            'lis_person_name_given':              'X',
-            'tool_consumer_instance_description': 'X',
-            'tool_consumer_instance_guid':        'elearning.u-pem.fr',
-            'oauth_consumer_key':                 'unknown',
-            'oauth_signature_method':             'HMAC-SHA1',
-            'oauth_timestamp':                    str(oauth2.generate_timestamp()),
-            'oauth_nonce':                        oauth2.generate_nonce(),
-            'roles':                              "Learner"
-        }
-        
-        norm_params = oauth_signature.normalize_parameters([(k, v) for k, v in params.items()])
-        # Last 'dns' is the arg use in request factory
-        uri = oauth_signature.normalize_base_string_uri("https://testserver/wims/1/")
-        base_string = oauth_signature.construct_base_string("POST", uri, norm_params)
-        
-        params['oauth_signature'] = oauth_signature.sign_hmac_sha1(base_string, "secret2", None)
-        request = RequestFactory().post(reverse("wims:wims_class", args=[1]), secure=True)
-        request.POST = params
-        
-        with self.assertRaises(BadRequestException) as r:
-            utils.is_valid_request(request)
-        self.assertEqual("Could not get a secret for key 'unknown'", str(r.exception))
-    
-    
-    def test_is_valid_request_timestamp_outdated(self):
-        params = {
-            'lti_message_type':                   'basic-lti-launch-request',
-            'lti_version':                        'LTI-1p0',
-            'launch_presentation_locale':         'fr-FR',
-            'resource_link_id':                   'X',
-            'context_id':                         'X',
-            'context_title':                      "A title",
-            'user_id':                            'X',
-            'lis_person_contact_email_primary':   'X',
-            'lis_person_name_family':             'X',
-            'lis_person_name_given':              'X',
-            'tool_consumer_instance_description': 'X',
-            'tool_consumer_instance_guid':        'elearning.u-pem.fr',
-            'oauth_consumer_key':                 'provider1',
-            'oauth_signature_method':             'HMAC-SHA1',
-            'oauth_timestamp':                    str(int(time.time()) - 2000),
-            'oauth_nonce':                        oauth2.generate_nonce(),
-            'roles':                              "Learner"
-        }
-        
-        norm_params = oauth_signature.normalize_parameters([(k, v) for k, v in params.items()])
-        # Last 'dns' is the arg use in request factory
-        uri = oauth_signature.normalize_base_string_uri("https://testserver/wims/1/")
-        base_string = oauth_signature.construct_base_string("POST", uri, norm_params)
-        
-        params['oauth_signature'] = oauth_signature.sign_hmac_sha1(base_string, "secret1", None)
-        request = RequestFactory().post(reverse("wims:wims_class", args=[1]), secure=True)
-        request.POST = params
-        
-        with self.assertRaises(PermissionDenied):
-            utils.is_valid_request(request)
-
-
-
-class CheckParametersTestCase(TestCase):
-    
-    def test_check_parameters_ok(self):
-        params = {
-            'lti_message_type':                   'basic-lti-launch-request',
-            'lti_version':                        'LTI-1p0',
-            'launch_presentation_locale':         'fr-FR',
-            'resource_link_id':                   'X',
-            'context_id':                         'X',
-            'context_title':                      "A title",
-            'user_id':                            'X',
-            'lis_person_contact_email_primary':   'X',
-            'lis_person_name_family':             'X',
-            'lis_person_name_given':              'X',
-            'tool_consumer_instance_description': 'X',
-            'tool_consumer_instance_guid':        'elearning.u-pem.fr',
-            'oauth_consumer_key':                 'provider1',
-            'oauth_signature_method':             'HMAC-SHA1',
-            'oauth_timestamp':                    str(oauth2.generate_timestamp()),
-            'oauth_nonce':                        oauth2.generate_nonce(),
-            'oauth_signature':                    oauth2.generate_nonce(),
-            'roles':                              "Learner"
-        }
-        params = utils.parse_parameters(params)
-        
-        self.assertIsNone(utils.check_parameters(params))
-    
-    
-    def test_check_parameters_custom_custom(self):
-        params = {
-            'lti_message_type':                   'basic-lti-launch-request',
-            'lti_version':                        'LTI-1p0',
-            'launch_presentation_locale':         'fr-FR',
-            'resource_link_id':                   'X',
-            'context_id':                         'X',
-            'context_title':                      "A title",
-            'user_id':                            'X',
-            'lis_person_contact_email_primary':   'X',
-            'lis_person_name_family':             'X',
-            'lis_person_name_given':              'X',
-            'tool_consumer_instance_description': 'X',
-            'tool_consumer_instance_guid':        'elearning.u-pem.fr',
-            'oauth_consumer_key':                 'provider1',
-            'oauth_signature_method':             'HMAC-SHA1',
-            'oauth_timestamp':                    str(oauth2.generate_timestamp()),
-            'oauth_nonce':                        oauth2.generate_nonce(),
-            'oauth_signature':                    oauth2.generate_nonce(),
-            'roles':                              "Learner",
-            'custom_custom_class_name':           "Custom",
-            'custom_custom_class_institution':    "Custom",
-        }
-        
-        with self.assertRaises(BadRequestException) as e:
-            utils.parse_parameters(params)
-        self.assertIn('custom_custom_class_name', str(e.exception))
-        self.assertIn('custom_custom_class_institution', str(e.exception))
-    
-    
-    def test_check_parameters_missing_lti_mandatory(self):
-        params = {
-            'lti_version':                        'LTI-1p0',
-            'launch_presentation_locale':         'fr-FR',
-            'resource_link_id':                   'X',
-            'context_id':                         'X',
-            'context_title':                      "A title",
-            'user_id':                            'X',
-            'lis_person_contact_email_primary':   'X',
-            'lis_person_name_family':             'X',
-            'lis_person_name_given':              'X',
-            'tool_consumer_instance_description': 'X',
-            'tool_consumer_instance_guid':        'elearning.u-pem.fr',
-            'oauth_consumer_key':                 'provider1',
-            'oauth_signature_method':             'HMAC-SHA1',
-            'oauth_timestamp':                    str(oauth2.generate_timestamp()),
-            'oauth_nonce':                        oauth2.generate_nonce(),
-            'oauth_signature':                    oauth2.generate_nonce(),
-            'roles':                              "Learner"
-        }
-        params = utils.parse_parameters(params)
-        
-        with self.assertRaises(BadRequestException) as r:
-            utils.check_parameters(params)
-        self.assertEqual("LTI request is invalid, missing parameter(s): ['lti_message_type']",
-                         str(r.exception))
-    
-    
-    def test_check_parameters_wimslti_mandatory(self):
-        params = {
-            'lti_message_type':                   'basic-lti-launch-request',
-            'lti_version':                        'LTI-1p0',
-            'launch_presentation_locale':         'fr-FR',
-            'resource_link_id':                   'X',
-            'context_id':                         'X',
-            'context_title':                      "A title",
-            'lis_person_contact_email_primary':   'X',
-            'lis_person_name_family':             'X',
-            'lis_person_name_given':              'X',
-            'tool_consumer_instance_description': 'X',
-            'tool_consumer_instance_guid':        'elearning.u-pem.fr',
-            'oauth_consumer_key':                 'provider1',
-            'oauth_signature_method':             'HMAC-SHA1',
-            'oauth_timestamp':                    str(oauth2.generate_timestamp()),
-            'oauth_nonce':                        oauth2.generate_nonce(),
-            'oauth_signature':                    oauth2.generate_nonce(),
-            'roles':                              "Learner"
-        }
-        params = utils.parse_parameters(params)
-        
-        with self.assertRaises(BadRequestException) as r:
-            utils.check_parameters(params)
-        self.assertEqual("LTI request is invalid, WIMS LTI require parameter(s): ['user_id']",
-                         str(r.exception))
 
 
 
@@ -387,7 +52,7 @@ class GetOrCreateClassTestCase(TestCase):
             'oauth_nonce':                        oauth2.generate_nonce(),
             'roles':                              settings.ROLES_ALLOWED_CREATE_WIMS_CLASS[0].value,
         }
-        params = utils.parse_parameters(params)
+        params = parse_parameters(params)
         
         wims = WIMS.objects.create(url="https://wims.u-pem.fr/",
                                    name="WIMS UPEM",
@@ -428,7 +93,7 @@ class GetOrCreateClassTestCase(TestCase):
             'oauth_nonce':                        oauth2.generate_nonce(),
             'roles':                              settings.ROLES_ALLOWED_CREATE_WIMS_CLASS[0].value,
         }
-        params = utils.parse_parameters(params)
+        params = parse_parameters(params)
         
         wims = WIMS.objects.create(url="https://wims.u-pem.fr/",
                                    name="WIMS UPEM", expiration=timedelta(days=400),
@@ -463,7 +128,7 @@ class GetOrCreateClassTestCase(TestCase):
             'oauth_nonce':                        oauth2.generate_nonce(),
             'roles':                              settings.ROLES_ALLOWED_CREATE_WIMS_CLASS[0].value,
         }
-        params = utils.parse_parameters(params)
+        params = parse_parameters(params)
         
         wims = WIMS.objects.create(url="https://wims.u-pem.fr/",
                                    name="WIMS UPEM", class_limit=123,
@@ -507,7 +172,7 @@ class GetOrCreateClassTestCase(TestCase):
             'custom_supervisor_lastname':         "Custom lastname",
             'custom_supervisor_firstname':        "Custom firstname",
         }
-        params = utils.parse_parameters(params)
+        params = parse_parameters(params)
         
         wims = WIMS.objects.create(url="https://wims.u-pem.fr/",
                                    name="WIMS UPEM",
@@ -559,7 +224,7 @@ class GetOrCreateClassTestCase(TestCase):
             'custom_supervisor_lastname':         "Custom lastname",
             'custom_supervisor_firstname':        "Custom firstname",
         }
-        params = utils.parse_parameters(params)
+        params = parse_parameters(params)
         wims = WIMS.objects.create(url="https://wims.u-pem.fr/", name="WIMS UPEM",
                                    ident="X", passwd="X", rclass="myclass")
         lms = LMS.objects.create(uuid="elearning.upem.fr", url="https://elearning.u-pem.fr/",
@@ -587,7 +252,7 @@ class GetOrCreateClassTestCase(TestCase):
             'roles':                              settings.ROLES_ALLOWED_CREATE_WIMS_CLASS[0].value,
             'custom_clone_class':                 wclass.qclass,
         }
-        params = utils.parse_parameters(params)
+        params = parse_parameters(params)
         _, wclass2 = utils.get_or_create_class(lms, wims, api, params)
         
         self.assertEqual(wclass.name, wclass2.name)
@@ -617,7 +282,7 @@ class GetOrCreateClassTestCase(TestCase):
             'oauth_nonce':                        oauth2.generate_nonce(),
             'roles':                              "None",
         }
-        params = utils.parse_parameters(params)
+        params = parse_parameters(params)
         
         wims = WIMS.objects.create(url="https://wims.u-pem.fr/",
                                    name="WIMS UPEM",
@@ -652,7 +317,7 @@ class GetOrCreateClassTestCase(TestCase):
             'oauth_nonce':                        oauth2.generate_nonce(),
             'roles':                              settings.ROLES_ALLOWED_CREATE_WIMS_CLASS[0].value,
         }
-        params = utils.parse_parameters(params)
+        params = parse_parameters(params)
         
         wims = WIMS.objects.create(url="https://wims.u-pem.fr/",
                                    name="WIMS UPEM",
@@ -660,10 +325,10 @@ class GetOrCreateClassTestCase(TestCase):
         lms = LMS.objects.create(uuid="elearning.upem.fr", url="https://elearning.u-pem.fr/",
                                  name="Moodle UPEM")
         wclass_db1 = WimsClass.objects.create(lms=lms, lms_uuid="77777", wims=wims,
-                                              wims_uuid="60001")
+                                              qclass="60001")
         supervisor = User("supervisor", "Supervisor", "", "password", "test@email.com")
         wclass1 = Class(wims.rclass, "A title", "UPEM", "test@email.com", "password",
-                        supervisor, lang="fr", qclass=wclass_db1.wims_uuid)
+                        supervisor, lang="fr", qclass=wclass_db1.qclass)
         wclass1.save(WIMS_URL, "myself", "toto")
         api = WimsAPI(WIMS_URL, "myself", "toto")
         
@@ -672,7 +337,7 @@ class GetOrCreateClassTestCase(TestCase):
         self.assertEqual(wclass_db1.lms, wclass_db2.lms)
         self.assertEqual(wclass_db1.lms_uuid, wclass_db2.lms_uuid)
         self.assertEqual(wclass_db1.wims, wclass_db2.wims)
-        self.assertEqual(wclass_db1.wims_uuid, wclass_db2.wims_uuid)
+        self.assertEqual(wclass_db1.qclass, wclass_db2.qclass)
         
         self.assertEqual(wclass1.email, wclass2.email)
         self.assertEqual(wclass1.name, wclass2.name)
@@ -706,7 +371,7 @@ class GetOrCreateUserTestCase(TestCase):
             'oauth_nonce':                        oauth2.generate_nonce(),
             'roles':                              "None",
         }
-        params = utils.parse_parameters(params)
+        params = parse_parameters(params)
         
         wims = WIMS.objects.create(url="https://wims.u-pem.fr/",
                                    name="WIMS UPEM",
@@ -714,10 +379,10 @@ class GetOrCreateUserTestCase(TestCase):
         lms = LMS.objects.create(uuid="elearning.upem.fr", url="https://elearning.u-pem.fr/",
                                  name="Moodle UPEM")
         wclass_db = WimsClass.objects.create(lms=lms, lms_uuid="77777", wims=wims,
-                                             wims_uuid="60002")
+                                             qclass="60002")
         supervisor = User("supervisor", "Supervisor", "", "password", "test@email.com")
         wclass = Class(wims.rclass, "A title", "UPEM", "test@email.com", "password",
-                       supervisor, lang="fr", qclass=wclass_db.wims_uuid)
+                       supervisor, lang="fr", qclass=wclass_db.qclass)
         wclass.save(WIMS_URL, "myself", "toto")
         
         user_db, user = utils.get_or_create_user(lms, wclass_db, wclass, params)
@@ -752,7 +417,7 @@ class GetOrCreateUserTestCase(TestCase):
             'oauth_nonce':                        oauth2.generate_nonce(),
             'roles':                              "None",
         }
-        params = utils.parse_parameters(params)
+        params = parse_parameters(params)
         
         wims = WIMS.objects.create(url="https://wims.u-pem.fr/",
                                    name="WIMS UPEM",
@@ -760,11 +425,11 @@ class GetOrCreateUserTestCase(TestCase):
         lms = LMS.objects.create(uuid="elearning.upem.fr", url="https://elearning.u-pem.fr/",
                                  name="Moodle UPEM")
         wclass_db = WimsClass.objects.create(lms=lms, lms_uuid="77777", wims=wims,
-                                             wims_uuid="60003")
+                                             qclass="60003")
         WimsUser.objects.create(lms=lms, lms_uuid='66', wclass=wclass_db, quser="jdoe")
         supervisor = User("supervisor", "Supervisor", "", "password", "test@email.com")
         wclass = Class(wims.rclass, "A title", "UPEM", "test@email.com", "password",
-                       supervisor, lang="fr", qclass=wclass_db.wims_uuid)
+                       supervisor, lang="fr", qclass=wclass_db.qclass)
         wclass.save(WIMS_URL, "myself", "toto")
         
         lastname = params['lis_person_name_family']
@@ -805,7 +470,7 @@ class GetOrCreateUserTestCase(TestCase):
             'oauth_nonce':                        oauth2.generate_nonce(),
             'roles':                              "None",
         }
-        params = utils.parse_parameters(params)
+        params = parse_parameters(params)
         
         wims = WIMS.objects.create(url="https://wims.u-pem.fr/",
                                    name="WIMS UPEM",
@@ -813,12 +478,12 @@ class GetOrCreateUserTestCase(TestCase):
         lms = LMS.objects.create(uuid="elearning.upem.fr", url="https://elearning.u-pem.fr/",
                                  name="Moodle UPEM")
         wclass_db = WimsClass.objects.create(lms=lms, lms_uuid="77777", wims=wims,
-                                             wims_uuid="60004")
+                                             qclass="60004")
         WimsUser.objects.create(lms=lms, lms_uuid='66', wclass=wclass_db, quser="jdoe")
         WimsUser.objects.create(lms=lms, lms_uuid='67', wclass=wclass_db, quser="jdoe1")
         supervisor = User("supervisor", "Supervisor", "", "password", "test@email.com")
         wclass = Class(wims.rclass, "A title", "UPEM", "test@email.com", "password",
-                       supervisor, lang="fr", qclass=wclass_db.wims_uuid)
+                       supervisor, lang="fr", qclass=wclass_db.qclass)
         wclass.save(WIMS_URL, "myself", "toto")
         
         lastname = params['lis_person_name_family']
@@ -860,7 +525,7 @@ class GetOrCreateUserTestCase(TestCase):
             'oauth_nonce':                        oauth2.generate_nonce(),
             'roles':                              "None",
         }
-        params = utils.parse_parameters(params)
+        params = parse_parameters(params)
         
         wims = WIMS.objects.create(url="https://wims.u-pem.fr/",
                                    name="WIMS UPEM",
@@ -868,11 +533,11 @@ class GetOrCreateUserTestCase(TestCase):
         lms = LMS.objects.create(uuid="elearning.upem.fr", url="https://elearning.u-pem.fr/",
                                  name="Moodle UPEM")
         wclass_db = WimsClass.objects.create(lms=lms, lms_uuid="77777", wims=wims,
-                                             wims_uuid="60005")
+                                             qclass="60005")
         user_db1 = WimsUser.objects.create(lms=lms, lms_uuid='77', wclass=wclass_db, quser="jdoe")
         supervisor = User("supervisor", "Supervisor", "", "password", "test@email.com")
         wclass = Class(wims.rclass, "A title", "UPEM", "test@email.com", "password",
-                       supervisor, lang="fr", qclass=wclass_db.wims_uuid)
+                       supervisor, lang="fr", qclass=wclass_db.qclass)
         wclass.save(WIMS_URL, "myself", "toto")
         
         lastname = params['lis_person_name_family']
@@ -914,7 +579,7 @@ class GetOrCreateUserTestCase(TestCase):
             'oauth_nonce':                        oauth2.generate_nonce(),
             'roles':                              settings.ROLES_ALLOWED_CREATE_WIMS_CLASS[0].value,
         }
-        params = utils.parse_parameters(params)
+        params = parse_parameters(params)
         
         wims = WIMS.objects.create(url="https://wims.u-pem.fr/",
                                    name="WIMS UPEM",
@@ -922,11 +587,11 @@ class GetOrCreateUserTestCase(TestCase):
         lms = LMS.objects.create(uuid="elearning.upem.fr", url="https://elearning.u-pem.fr/",
                                  name="Moodle UPEM")
         wclass_db = WimsClass.objects.create(lms=lms, lms_uuid="77777", wims=wims,
-                                             wims_uuid="60006")
+                                             qclass="60006")
         WimsUser.objects.create(lms=lms, lms_uuid=None, wclass=wclass_db, quser="supervisor")
         supervisor = User("supervisor", "Supervisor", "", "password", "test@email.com")
         wclass = Class(wims.rclass, "A title", "UPEM", "test@email.com", "password",
-                       supervisor, lang="fr", qclass=wclass_db.wims_uuid)
+                       supervisor, lang="fr", qclass=wclass_db.qclass)
         wclass.save(WIMS_URL, "myself", "toto")
         
         user_db, user = utils.get_or_create_user(lms, wclass_db, wclass, params)
