+++ lti_app/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+
+class LtiAppConfig(AppConfig):
+    name = 'lti_app'
