+++ lti_app/urls.py
@@ -9,12 +9,13 @@
 
 from django.urls import path
 
-from wims import views
+from lti_app import views
 
 
-app_name = "wims"
+app_name = "lti"
 
 urlpatterns = [
     path('<int:wims_pk>/', views.wims_class, name="wims_class"),
-    path('<int:wims_pk>/<int:activity_pk>', views.wims_activity, name="wims_activity"),
+    path('<int:wims_pk>/<int:activity_pk>/', views.wims_activity, name="wims_activity"),
+    path('links/', views.wims_activity, name="wims_activity"),
 ]
